refactor: move protobuf conversion logic to generated_types
