+++ data_types/src/lib.rs
@@ -16,7 +16,6 @@ mod database_name;
 pub use database_name::*;
 pub mod database_rules;
 pub mod error;
-pub mod field_validation;
 pub mod http;
 pub mod job;
 pub mod names;
